Add templating root repo

+++ _internal/templater/src/example/mod.rs
@@ -2,16 +2,18 @@ pub mod meta;
 
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct Config {
     pub meta: meta::Meta,
     pub display: Display,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct Display {
     pub title: String,
     pub tutorial_url: Option<String>,
     pub preview_file: Option<String>,
-    pub overview_weight: Option<f64>,
+
+    /// A higher number will show it before other examples.
+    pub overview_weight: Option<i64>,
 }
